+++ cdlclient/__init__.py
@@ -18,5 +18,6 @@ from cdlclient.baseproxy import SimpleBaseProxy  # noqa: F401
 from cdlclient.remote import SimpleRemoteProxy  # noqa: F401
 
 __version__ = "0.7.0"
+__required_server_version__ = "0.10.0"
 __docurl__ = "https://cdlclient.readthedocs.io/en/latest/"
 __homeurl__ = "https://github.com/Codra-Ingenierie-Informatique/DataLabSimpleClient/"