+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## Version 0.7.0 ##
 
+DataLab Simple Client is fully compatible with **DataLab 0.10.0** and above.
+With older versions of the DataLab server, some features may not work.
+
 💥 Changes:
 
 * Added `toggle_auto_refresh` method to `SimpleRemoteProxy`
 * Added `toggle_show_titles` method to `SimpleRemoteProxy`
+* Remote client is now checking the server version and shows a warning message if
+  the server version may not be fully compatible with the client version.
 
 ## Version 0.6.0 ##